Add oatts module declaration and import
This commit adds a new file `declarations.d.ts` which declares the module "oatts". Additionally, in the `makeOatts.ts` file, the code now imports the `oatts` module using an asterisk. The previous line that required 'oatts' is commented out.

The changes enable the usage of the "oatts" module for generating tests based on OpenAPI specifications.