Add validateOAS3 function
This commit adds a new file `validateOAS3.ts` which contains the implementation of the `validateOAS3` function. This function takes an OpenAPIV3 document as input and validates it against the OAI OpenAPI Specification v3.0 schema using the `@cfworker/json-schema` library. The result of the validation is returned as a Promise<ValidationResult>.