fAdd @babel/preset-typescript to devDependencies
This commit adds the "@babel/preset-typescript" package to the devDependencies section of the package.json file. This package is required for TypeScript support in Babel.

Additionally, other dependencies such as "ts-node", "tsconfig-paths-jest", and "wireit" have been added to the devDependencies section. The wireit dependency has also been configured with a new "generate" command and a modified "test" command that runs Jest.

The existing scripts in the scripts section have been updated accordingly to use the wireit commands for generating code and running tests.