Add version validation and dynamic version assignment
This commit adds code to validate the version number in the package.json file and assigns it dynamically to the `version` variable. If the version is invalid or doesn't match the expected format, an error will be thrown. The previous hard-coded version value has been replaced with the dynamic assignment.